Fixed Make files and Python Script
Note, global variables with the main method throw global variables not defined errors despite only trying to print one of the many variables. If trying to change those variables, you need to declare them as global inside of the main method to alter. The fix in this case was to move all global variables inside of the main method.